return a non-zero exit code on download error
Ask curl to return a non-zero exit code (22) and not produce output, when a web server returns a failure code (HTTP status other than 200).  Prior to this change, if the web server returned an error and content (like details of the REST error in the body), curl would still write to the archive file, and puppet would merrily chug along.  The only thing that would eventually stop this was having checksums enabled... but not all repos provide the checksum files.